requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-core>=0.1.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pandas>=2.0.0
python-telegram-bot>=20.0
web3>=6.0.0
pydantic>=2.0.0
openai>=1.0.0 